+++ public/md/history.md
@@ -1,5 +1,6 @@
 ##History
 
+*   20161117, 12:15: Refactored the navbar into a class.
 *   20161117, 09:35: Moved AJAX-functions and navbar to separate files/modules.
 *   20161116, 16:05: Quick test fetching content from DB with AJAX.
 *   20161115, 14:55: Set up for ES6 in the browser with System.js and Traceur.